Lots more work on an adaptive Clenshaw-Curtis quadrature scheme.
Working but to a very low accuracy at the moment. The problem is that
the order does not seem to be able to go high enough. However previous tests
showed that order 20ish should have been fine, so this may be a bug.

If not the next step is to break up the domain. However it will be
difficult to do this and still share knots between source points
evaluations.

Possibly we should just use a standardised scheme and see how bad the performance
hit is. Alternatively use Newton-Coates (evenly spaced knots) which would
allow for easy refinement while still being a progressive scheme.