Some small improvements to general stuff.
Attempted to write a quadrature function to use QUADPACK
algorithms. However QUADPACK requires the function be input as a
function pointer. Making function pointers requires that the function
be known at compile time (i.e. explicitly written in the code). Since
we want to integrate ~ #source_nodes * #test_functions different
functions in each element this is completely unreasonable.

I attempted to change the GSL QUADPACK implementation to allow the use
of function objects but it didn't work. I think there would be a lot
of messing with internals needed.