Working order-adaptive routine: fill_in_be_contribution_adaptive_not_stored! Still needs lots of cleaning up to the code.
Adaptive routine that takes advantage of storable knots has not yet been fixed.

Need to implement a GL-Konrad type quadrature if we decide to use GL (to allow at least some reuse of nodes).

Need to set a lower order minimum so that most calculations only take a small
amount of time.