Added templating of my face element by DIM so that we can know the DIM at compile time in the face element class (this lets us use it as a template).
Templated quadrature scheme by DIM.

Moved relative error calculations to a seperate function.

Added a check for very small absolute errors to avoid problems with relative
errors seeming large because the overall value is very small.

Changed adaptivity to use a list of orders instead of getting them from the
quadrature function. Before we were using the order as an index which left lots
of empty values in the storage vectors. Now we can use the index of the list of
orders instead. Also this makes it easier to compare some things.

Added some debugging output.