Boundary element calculations: ============================================================
 Added corner contribution to boundary element matrix (not in a very general way
unfortuantely). Also commented out the point element stuff since it's not in use
at the moment.

Commented out the Jacobian calculations since they are not in use at the moment.

Added a check that n is a unit vector in dgreendn calculation in PARANOID.
Removed a factor of two that I think is wrong in the same calculation.

Fixed the sign of the dgreendn calculation (I had included the minus sign twice
in different places).

Flux conditions:
============================================================

Changed the flux boundary condition face elements to get the shape/test
derivatives from the bulk instead of trying (and failing) to calculate it
themselves. Also moved shape/test/derivatives calculation function in micromag
element to public instead of protected since we need it in the face elements
(which do not inherit).

Fixed the flux residual contribution to be + instead of -.

General:
============================================================

Also some messing with parameters.