Added decimal points to all values which should be doubles just in case (got caught out by this: 2/3 = 0 but 2.0/3.0 = 0.6666666).
Removed factor of 1/4*pi - turned out it WAS already in the Lindholm formula.

Renamed a lot of variables interpolated -> itp, makes it much easier to
read/write code using them.

Started adding in Jacobian calculations (but got in a bit of a mess because of
the structuring of various for loops - this will be fixed next commit).

Sphere_driver
========================================

Pinned the value of phi_1 at node at [0 0 1] to zero to remove the additive
constant from the flux Poisson problem to allow comparison with an exact
solution. Does not affect the final solution since only the gradient of phi is
used.

A minus sign was not included in the Lindholm formula but needed for my formulation: added it to the get_boundary_matrix function (which is probably not the best place...).

Added some debugging code for comparison with exact solution.

Jacobian_algebra.m
========================================

Symbolic calculations for the Jacobian. Fairly rough code but works. Doesn't
give all components - see notebook 10/5/2012 for full calculations.