Got everything working with my Jacobian.
I was multiplying by the test function an extra time in the component of the m
residual due to exchange. Fixed this now

Changed the sign of the precession part of the LLG residual to match the proper
sign for the Gilbert formulation.

Moved some calculations outside of test function loop in field calculations.

Added (crude) adaptive timestepping to sphere_driver.

Added debugging switches to sphere_driver to turn on/off adaptive timestepping
and to easily swap back to a 2d model for faster calculations.