Added SumOfMatrices class and some tests.
Added SumOfMatrices class to deal with a summation of multiple different types
of matrix. This allows us to use appropriate methods for each matrix,
e.g. multiplication for dense and CR matrices uses the multiplication for those
classes.

Added test code for this new class.

Added fix to CRDoubleMatrix::clear - do not try to clean up memory for the
linear solver if non has been added.

Added easier to use constructor for CRDoubleMatrix from index notation.

Rename some variables for clarity in adaptive Newton solve.

Possible fix in the default preconditioner - it used to require an
OomphCommunicator* from the problem even though it was possible to call it with
no problem_pt. Changed to generate a dummy communicator_pt.