Fixed SumOfMatrices class.
Changed sum of matrix tests to use a sparse main matrix with a dense sub-block
instead of the other way around. This is much closer to my real use for the
class.

Uncovered a bug in sum of matrices - the map from the added matrix
multiplication result back to the full matrix was not done properly.

Fixed error checking to allow maps from main to added matrix with size less than
the number of rows/cols of the added matrix (useful for sparse matrices).