Added BEM to Jacobian.
Instead of pinning phi on the boundary we now insert zeros into the Jacobian for
those values. This is necessary because pinning a phi value results in the
removal of that equation from the Jacobian but we later need to add the
appropriate BEM elements for that equation to the Jacobian.

Overloaded get_jacobian in the problem to combine the sparse FEM jacobian with
our BEM Jacobian.

We now generate global equation to BEM equation maps based on phi and phi_1
equation numbers to give the appropriate positioning of the boundary matrix in
the Jacobian.

Added options for GMRES, mid-point method.

Split the main mesh into seperate bulk and flux meshes ready for use with
preconditioner framework.

(Re-)Added the adaptive integration scheme in case it is needed (e.g. for 2d
tests).