Much improved generic hybrid micromagnetics problem class.
We now use MagneticParamters class to pass all magnetic parameters in to
elements. This allowed many function pointers (and related functions) to be
removed.

Made a "finish_building" function to do all the stuff needed for BEM, flux
boundary conditions and other stuff needed for all magnetics problems.

Changed mumag4_driver.cc to work with this new class. It builds and runs but
does not converge, possibly because the sharp corners are not yet accounted for.