Remove factor of 2 in time part of exact wave-like solution
There is an additional factor of two in the exact solution of one of
Jeong's papers (as compared to other places where the exact solution can
be found). It seems that this factor is wrong (at least without some
kind of additional forcing): the numerical solution matches well without
this factor but not with it.